Refactor FetchRoleID, added fallback search of all guilds
If it can't find a named role in the config, it searches the currently requested 'guild'. If not found there, it searches the main guild. If still not found, it will search every configured guild.

DOWNSIDE:

Possible security flaw example:
Role "Manager" renamed in Main server to "XXRP | Manager" but not updated in config.
Role "Manager" exists for the manager(s) of a department.
Search fallbacks will give them perms / access in some cases.