Added error code 204
When calling the SetNickname export the code would throw an issue with the error code list. Upon further investigation I found that the SetNickname export would throw a member.code of 204. On a google search I found that error code 204 is "(No Content) status code indicates that the server has successfully fulfilled the request and that there is no additional content to send in the response payload body"

I modified the member.code check to ignore the code 204 and added code 204 to the list of error codes just in case