Rework GetDiscordRoles for multiguild compatibility
Now the GetDiscordRoles function will, by default, return a table containing the user's roles from all guilds they're a member of. However, if a guild is specified in argument 2, it will only fetch the roles from that guild. If multiguild option is disabled, it returns roles in main guild only.